refactor: move components to isolated directories
